requirements.txt
jupyter
numpy
pandas
black
flake8
nltk
scikit-learn
tf-models-official
tensorflow-text
tensorflow-hub
seaborn
